+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improved `pilet debug` to use actual app externals
 - Fixed `publicPath` in pilets built with Webpack to be `./` instead of `/`
+- Fixed default imports of shared dependencies with `v2` schema (#433)
 - Removed `file-loader` from `piral-cli-webpack5`
 
 ## 0.14.4 (December 1, 2021)