+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.14.5 (tbd)
 
 - Improved `pilet debug` to use actual app externals
+- Improved `piral debug` to respect `--public-url` like `piral build`
 - Fixed `publicPath` in pilets built with Webpack to be `./` instead of `/`
 - Fixed default imports of shared dependencies with `v2` schema (#433)
 - Fixed usage of custom public path with `piral-ng` (#434)